Push up current status
