+++ examples/LSTM/Project.toml
@@ -1,0 +1,13 @@
+[deps]
+DiffRules = "b552c78f-8df3-52c6-915a-8e097449b14b"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
+IRTools = "7869d1d1-7146-5819-86e3-90919afe41df"
+Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
+Metalhead = "dbeba491-748d-5e0e-a39e-b530a07fa0cc"
+Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
+NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+TensorFlow = "1d978283-2c37-5f34-9a8e-e9c0ece82495"
+Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
